+++ CHANGELOG.md
@@ -3060,3 +3060,4 @@
 - [2026-07-23 11:54] feat(i18n): add async i18n stream provider
 - [2026-07-23 12:00] refactor(network): extract reusable atomic widgets from network
 - [2026-07-23 12:21] refactor(i18n): extract reusable atomic widgets from i18n
+- [2026-07-23 13:17] feat(theme): implement responsive theme view component